Enable GITHUB_COM_TOKEN environment variable
When running this action on a private GitHub Enterprise
with self-hosted runners and you own analyse a golang project
you quickly run into GitHub.com rate-limiting.

According to the docs (https://docs.renovatebot.com/self-hosting/#githubcom-token-for-release-notes)
we need to add the environment variable `GITHUB_COM_TOKEN` to mitigate this.

This was not yet picked up by the docker input parser. This commit will enable `GITHUB_COM_TOKEN` 
to be used for Renovate.