feat: add mount-docker-socket option
Default to false. If set to true the action will mount the Docker socket
inside the renovate container so that the commands can use Docker. It
will also add the user inside of the renovate container to the Docker
group so that it has the right permissions to connect to the socket.