ci: ensure `example/` passes `renovate-config-validator`
As noted in #910, we have invalid configuration shared as an example,
which isn't ideal.

We should validate any JSON and `config.js` files in the example
directory.

We need to make sure that we override the ESM settings in this repo, as
otherwise we receive:

  module is not defined in ES module scope

When trying to validate the `renovate-config.js`.

This also follows the usage of a container from Renovate[0] for using
the container to run the validator.

[0]: https://github.com/renovatebot/.github/blob/2ab0f77818/.github/workflows/test.yml#L44